src/requirements.txt
anyio==3.6.2
asyncpg==0.27.0
bcrypt==4.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
ecdsa==0.18.0
exceptiongroup==1.0.1
fastapi==0.86.0
greenlet==2.0.0.post0
h11==0.14.0
idna==3.4
inflect==6.0.2
iniconfig==1.1.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pydantic==1.10.2
pyparsing==3.0.9
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlacodegen==2.3.0
SQLAlchemy==1.4.43
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
